refactor: use token to authorization
- 주문, 반품 토큰을 통한 요청 방식 적용

+++ src/components/exchange/FranchiseExchangeDetail.vue
@@ -72,6 +72,9 @@
 <script setup>
 import { ref } from "vue";
 import axios from 'axios';
+import FranchiseExchangePage from "./FranchiseExchangePage.vue";
+import { useStore } from 'vuex'; // Vuex store 임포트
+const store = useStore(); // Vuex store 사용
 
 const headers = ref([
   { key: 'productName', label: '상품명' },
@@ -97,11 +100,15 @@ const franchiseOwnerCode = ref(1);
 
 const deleteExchange = async () => {
   try {
+    const accessToken = store.state.accessToken;
+    if (!accessToken) {
+      throw new Error('No access token found');
+    }
     const response = await fetch(`/api/franchise/exchange/${item.exchangeCode}`, {
       method: 'DELETE',
       headers: {
-        "Content-Type": "application/json",
-        'access': `${localStorage.getItem('access')}`, // 인증 토큰을 포함하는 경우
+        'Authorization': `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
       },
       });
     if (response.status == 406) {
@@ -113,6 +120,7 @@ const deleteExchange = async () => {
       throw new Error('네트워크 오류 발생');
     }
     alert("삭제되었습니다.");
+    location.reload(FranchiseExchangePage);
     props.showDetailPopup();
   } catch (error) {
     console.error('오류 발생:', error);
